refactor: isolate source SDK development bootstrap
