+++ src/cli/commands/development.ts
@@ -2,9 +2,8 @@ import { createHash, randomUUID } from 'node:crypto';
 import { execFileSync, spawn, spawnSync } from 'node:child_process';
 import { closeSync, existsSync, lstatSync, mkdirSync, openSync, readFileSync, readlinkSync, readdirSync, renameSync, rmSync, statSync, symlinkSync, writeFileSync } from 'node:fs';
 import { basename, dirname, isAbsolute, relative, resolve } from 'node:path';
-import { fileURLToPath, pathToFileURL } from 'node:url';
-import { developmentCandidateSchema, developmentRuntimeSchema, type DevelopmentRuntime, type DevelopmentTarget } from '@treeseed/sdk/development';
-import { parse as parseYaml } from 'yaml';
+import { fileURLToPath } from 'node:url';
+import { developmentCandidateSchema, type DevelopmentRuntime, type DevelopmentTarget } from '@treeseed/sdk/development';
 import type { CommandContext, ParsedInvocation } from '../types.js';
 import { invokeLocalHostManager } from '../support/host-client.js';
 import { developmentStateRoot, selectDevelopmentCli } from './development-cli-selection.js';
@@ -15,6 +14,7 @@ import { applyDevelopmentRecovery, planDevelopmentRecovery } from './development
 import { ownsDevelopmentProcess, processIdentity } from './development-support/process-identity.js';
 import { developmentBootOrder } from './development-support/boot-order.js';
 import { assertNoSelectedDevelopmentCustody, managedContainerAlreadyReady, withDevelopmentLifecycle } from './development-support/lifecycle.js';
+import { loadDevelopmentRuntimes } from './development-support/runtime-loader.js';
 export { relativeOverlayTarget, startPackageSynchronizer, stopProcess, waitForNewPackageOverlay } from './development-support/overlays.js';
 export { developmentCliEntrypointPath, selectDevelopmentCli } from './development-cli-selection.js';
 
@@ -27,7 +27,6 @@ interface LocalSessionState {
 	candidates: string[];
 }
 
-interface ProjectSelection { manifest: string; worktree?: string; targets?: Array<{ id: string; mode: 'released' | 'candidate' | 'live' }> }
 interface DevelopmentStatusRecord {
 	session: {
 		targets: Array<{ projectId: string; targetId: string; mode: 'released' | 'candidate' | 'live'; generation: number; health?: string }>;
@@ -67,47 +66,6 @@ function repositoryClosure(runtime: DevelopmentRuntime, worktree: string, exclud
 	return { projectId: runtime.project.id, repository: runtime.project.repository, worktree, commit: git(worktree, ['rev-parse', 'HEAD']), branch, dirty: Boolean(status), dirtyDigest: status ? sha256(`${status}\n${git(worktree, ['diff', '--binary', 'HEAD', ...pathspec])}`) : null, recipeDigest: sha256(JSON.stringify(runtime)) };
 }
 
-function projectSelections(file: string): ProjectSelection[] {
-	const root = dirname(file), document = parseYaml(readFileSync(file, 'utf8')) as Record<string, unknown>;
-	if (document.development) return [{ manifest: file, worktree: root }];
-	if (!Array.isArray(document.projects) || document.projects.length === 0) throw new Error('Development session manifest requires a non-empty projects array.');
-	return document.projects.map((project) => {
-		if (!project || typeof project !== 'object' || Array.isArray(project)) throw new Error('Development project selection must be an object.');
-		const input = project as Record<string, unknown>;
-		if (typeof input.manifest !== 'string') throw new Error('Development project selection requires a manifest path.');
-		const manifest = isAbsolute(input.manifest) ? input.manifest : resolve(root, input.manifest);
-		const worktree = typeof input.worktree === 'string' ? (isAbsolute(input.worktree) ? input.worktree : resolve(root, input.worktree)) : dirname(manifest);
-		return { manifest, worktree, ...(Array.isArray(input.targets) ? { targets: input.targets as ProjectSelection['targets'] } : {}) };
-	});
-}
-
-async function localDevelopmentRuntimeSchema(selections: ProjectSelection[], prepare: boolean) {
-	const sdk = selections.find((selection) => {
-		const document = parseYaml(readFileSync(selection.manifest, 'utf8')) as { development?: { project?: { id?: unknown; repository?: unknown } } };
-		return document.development?.project?.id === 'sdk' && document.development.project.repository === 'treeseed-ai/sdk';
-	});
-	if (!sdk) return developmentRuntimeSchema;
-	const worktree = sdk.worktree ?? dirname(sdk.manifest), packagePath = resolve(worktree, 'package.json');
-	const packageDocument = JSON.parse(readFileSync(packagePath, 'utf8')) as { name?: unknown; scripts?: { ['build:dist']?: unknown } };
-	if (packageDocument.name !== '@treeseed/sdk' || typeof packageDocument.scripts?.['build:dist'] !== 'string') throw new Error('The selected SDK source is not a buildable @treeseed/sdk worktree.');
-	if (prepare) {
-		const result = spawnSync('npm', ['run', 'build:dist'], { cwd: worktree, env: process.env, stdio: 'inherit', timeout: 900_000 });
-		if (result.status !== 0) throw new Error('The selected SDK source contract could not be built.');
-	}
-	const entrypoint = resolve(worktree, 'dist/development/index.js');
-	if (!existsSync(entrypoint)) throw new Error('The selected SDK source contract is not built. Run development session start without --plan to prepare it.');
-	const module = await import(`${pathToFileURL(entrypoint).href}?source=${statSync(entrypoint).mtimeMs}`) as { developmentRuntimeSchema?: typeof developmentRuntimeSchema };
-	if (!module.developmentRuntimeSchema?.parse) throw new Error('The selected SDK source does not export its development runtime contract.');
-	return module.developmentRuntimeSchema;
-}
-
-async function loadRuntimes(file: string, prepareLocalSdk = false) {
-	const selections = projectSelections(file), schema = await localDevelopmentRuntimeSchema(selections, prepareLocalSdk);
-	return selections.map((selection) => {
-		const document = parseYaml(readFileSync(selection.manifest, 'utf8')) as { development?: unknown };
-		return { selection, runtime: schema.parse(document.development) };
-	});
-}
 
 function hostCommand(handlerId: string, payload: unknown) {
 	return { handlerId, arguments: [], options: { payload: JSON.stringify(payload) } };
@@ -210,7 +168,7 @@ async function waitForDirectReadiness(target: DevelopmentTarget, timeoutSeconds:
 async function startSession(invocation: ParsedInvocation, context: CommandContext) {
 	const manifest = resolve(context.cwd, invocation.arguments[0]!);
 	if (invocation.options.plan !== true) assertNoSelectedDevelopmentCustody(context.env);
-	const projects = await loadRuntimes(manifest, invocation.options.plan !== true);
+	const projects = await loadDevelopmentRuntimes(manifest, invocation.options.plan !== true);
 	const now = new Date();
 	const sessionId = `dev-${randomUUID().slice(0, 12)}`;
 	const targets = projects.flatMap(({ selection, runtime }) => (selection.targets ?? runtime.targets.map((target) => ({ id: target.id, mode: target.kind === 'rebuild-restart' ? 'candidate' as const : 'released' as const }))).map((target) => ({ projectId: runtime.project.id, targetId: target.id, mode: target.mode, generation: 0, health: target.mode === 'released' ? 'ready' as const : 'pending' as const })));
@@ -494,7 +452,7 @@ async function runDevelopmentUnlocked(invocation: ParsedInvocation, context: Com
 		if (invocation.options.plan === true) return { sessionId, restore: true, mutation: false };
 		const running = await stopProcesses(state);
 		const active = new Set(running.map((entry) => `${entry.projectId}.${entry.targetId}`));
-		for (const { selection, runtime } of await loadRuntimes(state.manifest)) for (const target of runtime.targets) {
+		for (const { selection, runtime } of await loadDevelopmentRuntimes(state.manifest)) for (const target of runtime.targets) {
 			if (usesManagedContainer(target)) await containerOperation(context, sessionId, runtime, target, 'stop');
 			else if (active.has(`${runtime.project.id}.${target.id}`) && target.operations.cleanup) runOneShotOperation(state, target.operations.cleanup, selection.worktree!, 'released', context.env, { TREESEED_DEVELOPMENT_CLEANUP_SCOPE: 'session' });
 		}
@@ -505,7 +463,7 @@ async function runDevelopmentUnlocked(invocation: ParsedInvocation, context: Com
 	if (invocation.command.name === 'dev logs') {
 		const selected = typeof invocation.options.target === 'string' ? invocation.options.target : null;
 		const logs: unknown[] = Object.entries(state.processes).filter(([key]) => !selected || key === selected).map(([target, processState]) => ({ target, path: processState.log, bytes: existsSync(processState.log) ? statSync(processState.log).size : 0 }));
-		for (const { runtime } of await loadRuntimes(state.manifest)) for (const target of runtime.targets) {
+		for (const { runtime } of await loadDevelopmentRuntimes(state.manifest)) for (const target of runtime.targets) {
 			const key = `${runtime.project.id}.${target.id}`;
 			if (usesManagedContainer(target) && (!selected || selected === key)) logs.push({ target: key, diagnostics: await containerOperation(context, sessionId, runtime, target, 'logs') });
 		}
